test: relocate finalize failure tests next to their validators
Move the four finalize snapshot tests out of parse.rs (which is scoped to
parse_config pipeline diagnostics) and place each next to the code that
actually raises the error:

- finalize_no_sources -> lifecycle.rs, where Config::finalize's NoSources
  check lives (new test module)
- finalize_base_path/route_prefix -> config.rs, next to parse_base_path
- finalize_postgres_missing_connection_string -> postgres/config.rs, next
  to PgConfig::finalize (drops the now-redundant postgres cfg gate)

The shared render_finalize_failure helper stays in test_helpers.rs.
