+++ martin/src/config/file/main/config.rs
@@ -259,6 +259,7 @@ mod tests {
     use super::*;
     use crate::MartinError;
     use crate::config::file::CachePolicy;
+    use crate::config::test_helpers::render_finalize_failure;
     use crate::logging::LogFormat;
 
     #[test]
@@ -289,6 +290,28 @@ mod tests {
     fn parse_base_path_rejects_invalid_paths() {
         parse_base_path("").unwrap_err();
         parse_base_path("foo/bar").unwrap_err();
+    }
+
+    #[test]
+    fn finalize_base_path_must_start_with_slash() {
+        insta::assert_snapshot!(
+            render_finalize_failure(indoc::indoc! {"
+                pmtiles: /tmp
+                base_path: not-a-path
+            "}),
+            @"Base path must be a valid URL path, and must begin with a '/' symbol, but is 'not-a-path'"
+        );
+    }
+
+    #[test]
+    fn finalize_route_prefix_must_start_with_slash() {
+        insta::assert_snapshot!(
+            render_finalize_failure(indoc::indoc! {"
+                pmtiles: /tmp
+                route_prefix: oops
+            "}),
+            @"Base path must be a valid URL path, and must begin with a '/' symbol, but is 'oops'"
+        );
     }
 
     #[test]
